feat: interefacesを追加。calendarのイベントが一部repository層を直接叩いていたので、usecase層から呼び出すように修正
